Main component update
- Add voice validation
- Add argument checking after asking(using for prevent errors and dissmiss action)
- Add select action

+++ main.py
@@ -1,31 +1,30 @@
 #!/usr/bin/python
 
-import voice
+from voice import Voice
 import keylogger
 from fs import FS
 
 
-""" Output for debugging """
-def onScreenOutput(out):
-	print out
-
+v = Voice()
 """ Mounting output to file system """
-fs = FS(onScreenOutput)
+fs = FS(v.voiceOutput)
 
+selectedItem = ''
 """ Main loop """
 while True:
 	key = keylogger.getkey()
 	
 	if key == 'l':
 		""" ls """
-		fs.scanDir()
-	elif key == 'd':
+		fs.getOutput()(fs.scanDir())
+	elif key == 'p':
 		""" pwd """
 		fs.getOutput()(fs.currentDir)
 	elif key == 'n':
 		""" cd """
 		folder = raw_input("Enter folder: ")
-		fs.navigate(folder)
+		if folder:
+			fs.navigate(folder)
 	elif key == 'm':
 		""" mkdir """
 		folder = raw_input("Enter name for folder: ")
@@ -33,21 +32,39 @@ while True:
 	elif key == 'o':
 		""" xdg-open or touch """
 		filename = raw_input("Enter filename: ")
-		if fs.isItemExists(filename):
-			fs.openFile(filename)
-		else:
-			fs.createFile(filename)
-			fs.editFile(filename)
+		if filename:
+			if fs.isItemExists(filename):
+				fs.openFile(filename)
+			else:
+				fs.createFile(filename)
+				fs.editFile(filename)
 	elif key == 'd':
 		""" rm """
 		filename = raw_input("Enter itemname: ")
-		fs.removeItem()
-	elif key == 'r':
-		""" mv(rename) """
+		if filename:
+			fs.removeItem()
+	elif key == 'x':
+		""" cut(for future paste) """
+		selection = v.selectFrom(fs.scanDir())
+		if selection:
+			selectedItem = fs.currentDir + selection
 	elif key == 'c':
 		""" copy """
+		if selectedItem:
+			fs.paste(selectedItem)
+			selectedItem = ''
+		else:
+			v.voiceOutput("Please select item first")
+	elif key == 'v':
+		""" paste """
+		if selectedItem:
+			fs.move(selectedItem)
+			selectedItem = ''
+		else:
+			v.voiceOutput("Please select item first")
 	elif key == 'i':
 		""" get properties """
+		
 	elif key == 'f':
 		""" find . -name """
 	else:
